review changes

+++ extensions/rocksdb-repos/ProvenanceRepository.h
@@ -162,7 +162,7 @@ class ProvenanceRepository : public core::Repository, public std::enable_shared_
       std::string key = it->key().ToString();
       if (store.size() >= max_size)
         break;
-      if (eventRead->DeSerialize(reinterpret_cast<const uint8_t *>(it->value().data()), static_cast<int>(it->value().size()))) {
+      if (eventRead->DeSerialize(reinterpret_cast<const uint8_t *>(it->value().data()), it->value().size())) {
         store.push_back(std::dynamic_pointer_cast<core::CoreComponent>(eventRead));
       }
     }
@@ -178,7 +178,7 @@ class ProvenanceRepository : public core::Repository, public std::enable_shared_
         break;
       std::shared_ptr<core::SerializableComponent> eventRead = lambda();
       std::string key = it->key().ToString();
-      if (eventRead->DeSerialize(reinterpret_cast<const uint8_t *>(it->value().data()), static_cast<int>(it->value().size()))) {
+      if (eventRead->DeSerialize(reinterpret_cast<const uint8_t *>(it->value().data()), it->value().size())) {
         max_size++;
         records.push_back(eventRead);
       }
@@ -194,7 +194,7 @@ class ProvenanceRepository : public core::Repository, public std::enable_shared_
       std::string key = it->key().ToString();
       if (records.size() >= (uint64_t)maxSize)
         break;
-      if (eventRead->DeSerialize(reinterpret_cast<const uint8_t *>(it->value().data()), static_cast<int>(it->value().size()))) {
+      if (eventRead->DeSerialize(reinterpret_cast<const uint8_t *>(it->value().data()), it->value().size())) {
         records.push_back(eventRead);
       }
     }
@@ -207,7 +207,7 @@ class ProvenanceRepository : public core::Repository, public std::enable_shared_
       std::shared_ptr<ProvenanceEventRecord> eventRead = std::make_shared<ProvenanceEventRecord>();
       std::string key = it->key().ToString();
 
-      if (store.at(max_size)->DeSerialize(reinterpret_cast<const uint8_t *>(it->value().data()), static_cast<int>(it->value().size()))) {
+      if (store.at(max_size)->DeSerialize(reinterpret_cast<const uint8_t *>(it->value().data()), it->value().size())) {
         max_size++;
       }
       if (store.size() >= max_size)
